+++ category/Foundation/NSFileManager+RFKit.h
@@ -12,7 +12,8 @@
 #import <Foundation/Foundation.h>
 
 @interface NSFileManager (RFKit)
-- (NSArray *)subDirectoryOfDirectoryAtPath:(NSString *)path error:(NSError *__autoreleasing *)erro;
+
+- (NSArray *)subDirectoryOfDirectoryAtPath:(NSString *)path error:(NSError *__autoreleasing *)error;
 
 /** Performs a deep search of the specified directory and returns URLs for the contained items which has given file extension.